user edit create 結合

+++ app/Http/Controllers/UserController.php
@@ -15,46 +15,39 @@ class UserController extends Controller
         $loginuser = Auth::user();
         return compact('loginuser');
     }
+    
     public function index()
     {
         return User::all();
     }
 
-    public function show(User $user)
-    {
-        return $user;
-    }
-
-    public function store(Request $request, User $user)
-    {
-        if(isset($request->file)){
-            $request->file->storeAs('public/', $request->img_name);
-            $user["img_name"] = $request->img_name;
-        }
-        $user["name"] = $request->name;
-        $user["email"] = $request->email;
-        $user["password"] = $request->password;
-        $user["salary"] = $request->salary;
-        $user["img_name"] = $request->img_name;
-        $user->save();
-
-        return $user;
-    }
     public function update(Request $request, User $user)
     {
-        if(isset($request->file)){
-            $request->file->storeAs('public/', $request->img_name);
-            Storage::delete('public/' . $request->img_oldname);
+        if(isset($request->id)){
+            if(isset($request->file)){
+                $request->file->storeAs('public/', $request->img_name);
+                Storage::delete('public/' . $request->img_oldname);
+                $user->where("id", $request->id)->update([
+                    "img_name" => $request->img_name,
+                ]);
+            }
             $user->where("id", $request->id)->update([
-                "img_name" => $request->img_name,
+                "name" => $request->name,
+                "email" => $request->email,
+                "salary" => $request->salary,
             ]);
-        }
-
-        $user->where("id", $request->id)->update([
-            "name" => $request->name,
-            "email" => $request->email,
-            "salary" => $request->salary,
-        ]);
+        }else{
+            if(isset($request->file)){
+                $request->file->storeAs('public/', $request->img_name);
+                $user["img_name"] = $request->img_name;
+            }
+            $user["name"] = $request->name;
+            $user["email"] = $request->email;
+            $user["password"] = $request->password;
+            $user["salary"] = $request->salary;
+            $user["img_name"] = $request->img_name;
+            $user->save();
+        };
     }
     public function destroy(User $user)
     {
